fixed update and delete methods for comments

+++ backend/CulturalHeritage/src/main/java/tim2/CulturalHeritage/service/CommentServiceImpl.java
@@ -1,13 +1,17 @@
 package tim2.CulturalHeritage.service;
 
+import java.security.AccessControlException;
+
 import javax.persistence.EntityNotFoundException;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import tim2.CulturalHeritage.model.AuthenticatedUser;
 import tim2.CulturalHeritage.model.Comment;
 import tim2.CulturalHeritage.model.FileDB;
 import tim2.CulturalHeritage.repository.CommentRepository;
@@ -39,17 +43,40 @@ public class CommentServiceImpl implements CommentService {
     }
 
     @Override
-    public Comment update(Comment comment, MultipartFile file) {
+    public Comment update(Comment comment, MultipartFile file) throws AccessControlException {
+
+        //if id is not found in the db
         if(null == commentRepository.findById(comment.getId()).orElse(null)){
             throw new EntityNotFoundException();
         }
+
+        //if one user is trying to update other user's comment
+        AuthenticatedUser userFromDB = commentRepository.findById(comment.getId()).orElse(null).getAuthenticatedUser();
+        if(comment.getAuthenticatedUser().getId() != userFromDB.getId()){
+            throw new AccessControlException("not allowed");
+        }
+
         FileDB fileDB = fileDBService.add(file);
         comment.setImages(fileDB);
         return commentRepository.save(comment);
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws AccessControlException {
+
+        //if id is not found in the db
+        if(null == commentRepository.findById(id).orElse(null)){
+            throw new EntityNotFoundException();
+        }
+
+
+        //if one user is trying to update other user's comment
+        AuthenticatedUser currentUser = (AuthenticatedUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        AuthenticatedUser userFromDB = commentRepository.findById(id).orElse(null).getAuthenticatedUser();
+        if(currentUser.getId() != userFromDB.getId()){
+            throw new AccessControlException("not allowed");
+        }
+
         commentRepository.deleteById(id);
     }
